DOP-3203: Rename associations to associated_products
